// lib/knowledge_main/types.ts
export interface Document {
  docuId: string
  name: string
  knowledgeType: string
  documentType: string
  departmentName: string
  creationYear: string
  creator: string
  projectName: string
  equipmentName: string
  referenceId: string
  storageLocation: string
  storageId: string
  level: string
  likeCount: number
  comment: string // コメント欄を追加
}